Revert "Handle general {QR,LQ}PackedQ in det(Q)"
This reverts commit f152120fd08b144dbad011c5d5bb1d282053ae33.

+++ stdlib/LinearAlgebra/test/qr.jl
@@ -249,7 +249,7 @@ end
 @testset "det(Q::QRPackedQ)" begin
     @testset for n in 1:3, m in 1:3
         Q, = qr(randn(n, m), Val(true))
-        @test det(Q) ≈ det(collect(Q))
+        @test det(Q)::Int ≈ det(collect(Q))
     end
 end
